+++ CHANGELOG.md
@@ -2,9 +2,6 @@
 
 ## 4.0
 
- - TODO: Prevent teleporting into an arena
- - TODO: Teleporting out of a mob arena broke it
-
  - Updated to Spigot API 1.14, will no longer work on previous versions. May also require Magic 8.1 or higher.
  - Add item_wear arena property, to turn off items and armor getting worn in a battle
  - Add allow_consuming arena property
@@ -12,6 +9,9 @@
  - Add leaderboard_sign_type parameter to arenas
  - Workaround problem with Spigot update that adds &r at the end of sign text
  - Add allow_melee and allow_projectiles flags, for making magic-only arenas
+ - Add arena stages:
+   - New `/arena stage` command to manage stages
+   - Mob spawns and rewards can be added to each stage
 
 ## 3.4